Added sysfs core's interface and governor callbacks
This patch add support for governor registration and FSC building callbacks.
A sysfs interface allows to enable/disable cpm.
Governor callbacks are used only when both cpm is enabled and a device is
added/removed.
The governor callback is executed within a workqueue in order to reduce
the latency for devices registration.

TODO
 - monitor device set changing while cpm is disabled and trigger a new FSC
   update once it will be enabled again.