CPM: Testing
This patch provide some testing support in order to simply evaluate some
subsystems interface and integration.
A dummy policy ir provided which simply return the un-ordered FSC list it
gets in input. A dummy driver is also provided to test the framework on
different scenarios: this module is usefaoul to verify the framework
overheads.