Added first perf counter binding
This patch add the required support to configure and enable/disable CPUs
performance counters. A single PMU counter is used so far, and it
statically reference the L1D cache misses. A configuration file is
foreseen as a useful and generic interface to define the PMU counter
of interest.

NOTE: The actual tracepoint is placed on a codepath where interrupts are
disabled. Thus, the perf_event_read_value result into a slowpath WARNING
because of a potential deadlock. However, we should investigate better
if the deadlock conditionc could happen considering the TP site (which
is at the beginning of a context switch), or find a better location
outside of the slow_path but still allowing to capture all task related
cache events.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>