Added initial multi-counters support
This patch make use of event arrays in order to support the usage of
multiple per-CPU event counters.

Arrays are statically allocated to match the tracepoint data.
An optimization could be to dynamically allocate CPU events, according to
user run-time needs, and pad the ring-buffer event counters (since the
tracepoint record could not be changed at run-time).
Pheraps this option is not good for performances?!

Signed-off-by: Patrick Bellasi <derkling@gmail.com>