ARM: Allow for more than 2 timers to be enabled by platform code.
The SP804 timer code has static structures for only two timers to use (one
as a free-running source and one as a timer event). Most platforms however
have at least 4 timers available, with the remaining two being unused due
to the static structure restrictions.

This change allows for dynamic allocation of timer structures. Note that
the clock framework has no method for removing the timers added from the
timers list, so timers that can appear and dissapear from the memory space
should not use this code.