ARM: VExpress - Tuscan: Remove VEXPRESS_PCIE_RC_IN_FPGA config option.
PCI-Express ROOT Complex for Tuscan has been initially developed on FPGA.
Kernel support for it has been added under VEXPRESS_PCIE_RC_IN_FPGA option.
As this is the only version we are now supporting for Tuscan, remove the
option and the code that is accessed when this config option is not selected.