ARM: kernel: save/restore multi CPU implementation
This patch adds the multi arch implementation of code required
to save/restore context for an ARM CPU.

Currently supported architecture is v7, inclusive of A5, A8 and A9
processors.

The CPU type is detected at boot, when the basic save/restore
structures are allocated. Function pointers are initialized and used
in a CPU independent manner through an API allowing to put a given CPU
(and cluster) in a given C state.

L2 retention is fully supported.

The current implementation just enables save/restore on Linux and
the API, called from CPUidle is flexible enough to choose the HW
configuration that must be saved/restored. The code is undergoing a
rework, where most of the saving/restoring of HW state is going to
be split and moved to the respective kernel subsystems (e.g VFP, GIC).

In order to make the code SoC independent, power controller commands
should be executed in the respective mach subsystems using a notifier
block, this is still work in progress.

Tested on ELBA SMP with single A9 CPU shutdown, woken up by the other
running CPU when it enters the idle loop.

Integration with proper firmware is ongoing.