ARM: vexpress: add support for mainboard CLDC
The V2M board has an on-board CLDC, together with some RAM.  This
device is usually left unused with running a ct-ca9x4 tile, but
becomes very useful with a ct-ca5s or ct-ca15x4, none of which
contains a CLCD block.

Add minimal support to enable this device.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>

Conflicts:

        arch/arm/mach-vexpress/v2m.c (trivial fixup)

Signed-off-by: Dave Martin <dave.martin@linaro.org>