ARM: vexpress: Gross extra hacks for mainboard CLCD on Fast Models
This patch hacks some addresses, clock lookup tables etc. to enable
use of the mainboard CLCD on the ARM Fast Models.  This is the
Wrong Way To Do It -- this patch needs to be dropped in favour of
doing things the proper DT way.  For now, this at least gets the
CLCD working.

Signed-off-by: Dave Martin <dave.martin@linaro.org>