ARM: columbus: CPU idle baseline implementation
This patch defines the skeleton driver for the complete Columbus
CPU idle implementation. All CPU states are supported inclusive of
cluster states.

In the Columbus platform the kernel runs in non-secure mode so that all
power down requests coming from OSPM subsystems have to be executed through
an SMC interface.

For that purpose, the patch also defines an assembly stub implementing the
CPU shutdown SMC call defined in the OSPM design document.

The driver keeps a cpumask of CPUs in deep shutdown so that the driver
is capable of eliciting the proper cluster state before issuing the
shutdown SMC call.

Idle code is heavily based on common save/restore, in particular the
driver relies on CPU PM notifiers to save/restore the state of VFP and
GIC, and calls into cpu_suspend in order to save the CPU context.

Currently, C-states are defined per-driver so that it is impossible for
the driver to cater for the C-state table differences that come up when
every CPU can be switched to a CPU in another cluster behind kernel
control.