ARM: columbus: add CPU idle debugfs entries for C-state enablement
This patch adds a simple debugfs interface to enable/disable deep
C-states for all CPUs in the columbus cluster.

It introduces a bitmap that can be programmed using a simple debugfs
entry accordingly, by echoing values in:

/sys/kernel/debug/idle_debug/enable_idle

[0-N]*                  enable deep C-states for the given CPU
[num_possible_cpus]     enable deep C-states for all CPUs and cluster
[0xff]                  disable C-states for all CPUs

(*) N = (num_possible_cpus - 1)

The current deep C-states enabled cpumask can be obtained by reading

/sys/kernel/debug/idle_debug/enable_mask

The output value is provided as a list of enabled CPUs.