cpufreq/arm-bl-cpufreq: Add simple cpufreq big.LITTLE switcher frontend
This patch adds a very simple cpufreq-based frontend to the ARM
big.LITTLE switcher.

This driver simply simulates two performance points corresponding
to the big and little clusters.

Note that this driver requires the ARM switcher implementation to
be loaded in order to work.  The switcher must have been built with
ASYNC = FALSE in big-little/Makefile in order for the synchronous
switching interface to be exposed.

Bugs and limitations:

  * Switching twice in quick succession currently tends to cause a
    deadlock inside the switcher.  I still need to identify exactly
    what is going wrong here.

  * There is currently no tracing interface, and no interface for
    reporting what the dummy frequencies exposed by the driver
    actually mean in terms of real cluster / performance point
    combinations.  For the very simple case supported,
    cpuinfo_max_freq corresponds to big and cpuinfo_min_freq
    corresponds to LITTLE.

  * cpufreq will trigger spurious extra cluster switches.  This is
    a "feature" since I didn't tell cpufreq that this matters, but
    it may not be desirable.  Setting policy->cpus can probably fix
    this.

Signed-off-by: Dave Martin <dave.martin@linaro.org>