cpufreq/arm-bl-cpufreq: cpufreq transition notifications support
This patch adds transition notifications using
cpufreq_notify_transition().  This is needed in order for the
expected values to appear when reading scaling_cur_freq and
scaling_setspeed in /sys/devices/system/cpu/*/cpufreq/.

To make this easier, some minor refactoring is done so that struct
cpufreq_frequency_table * is passed around instead of
frequencies/cluster numbers, where appropriate.  This makes it
easier to get the frquency and cluster values for a given
performance point.

Signed-off-by: Dave Martin <dave.martin@linaro.org>