ARM: gic: add CPU migration support
The git_migrate_target() changes the CPU interface mapping for the
current CPU to redirect SGIs to the given interface, and it also updates
the target CPU for each interrupts to the provided CPU interface if
they were targeting the current interface.

This must be called before gic_cpu_save() so no new SGIs are routed to
the current interface while the already pending ones are preserved and
eventually restored on the new interface via gic_cpu_restore().

Signed-off-by: Nicolas Pitre <nico@linaro.org>