ARM: Columbus: Add DT support for reading CPU frequencies
This patch adds support for parsing and reading the device tree
to get the CPU operating points(OPP). Currently only frequency
support is added but can be easily extended to voltages.
It tries to parse the data from FDT for CPU frequencies if available.
In case of failure, it continues to probe SCP to obtain OPPs

Signed-off-by: Sudeep KarkadaNagesha <Sudeep.KarkadaNagesha@arm.com>
Reviewed-by: Liviu Dudau <Liviu.Dudau@arm.com>