cpufreq/arm-bl-cpufreq: remove the forced switch on module removal
Switching all CPUs back to the fast cluster upon driver removal
might be considered a good thing.  However this has some
disadvantages such as not allowing for a specific power state to be
left active once the driver is removed.  Codewise, special care
would also be needed for this case as the notifiers can't be called
once the driver is unregistered (the code is buggy in that regard
already, oopsing the kernel).

Removing this driver is rather unlikely in practice, and people
doing so should know what they're doing.  So the best solution is
probably to just remove the forced switch from the module exit
path.

(Trivially ported over to the reference switcher driver by Dave
Martin.)

Signed-off-by: nicolas Pitre <nico@linaro.org>
Signed-off-by: Dave Martin <dave.martin@linaro.org>