cpufreq/arm-bl-cpufreq: Provide fake transition latency value
Some governors refuse to work unless a cpufreq driver advertises a
finite transition latency value.

We can't set this to something sensible until we have real hardware
to calibrate on.  In the meantime, we can define this to an
arbitrary value to permit experimentation with governors such as
conservative and ondemand.

Signed-off-by: Dave Martin <dave.martin@linaro.org>