ARM: arch_timers: enable the use of the virtual timer
At the moment, the arch_timer driver only uses the physical timer,
which can cause problem if PL2 hasn't enabled PL1 access in CNTHCTL,
which is likely in a virtualized environment. Instead, the virtual
timer is always available.

This patch enables the use of the virtual timer, unless no
interrupt is provided in the DT for it, in which case it falls
back to the physical timer.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>

Conflicts:

	arch/arm/kernel/arch_timer.c

Change-Id: Ib8aa8b833e9758d7d8fcd705c58bcc31b938b1b7