ARM: b.L switcher tests: port existing tests from reference switcher
Port in-modules tests on arm-bl-cpufreq cpufreq driver from reference switchers.

Tests are mainly the same as on reference switcher, we just test all A7-A15 pair separately and not all together.

To build the tests, add the following config option:

ARM_BL_CPUFREQ_TEST=y

The tests will run when the driver is loaded, but they are not run
by default.  To run the tests, you need to pass the option
test_config=1 to the arm-bl-cpufreq driver when loading it.

Signed-off-by: Mathieu Briand <mathieu.briand@linaro.org>
Signed-off-by: Dave Martin <dave.martin@linaro.org>
Signed-off-by: Nicolas Pitre <nicolas.pitre@linaro.org>