ARM: formalize an IPI for CPU wake-ups
Many CPUs are implementing power saving states via the WFI mode.
In some cases, WFI must be entered in order to suspend, or even reset,
a CPU.  In those cases it is possible to cancel a reset state before
it is effective, however the CPU must still receive an interrupt to exit
WFI.

Let's formalize an IPI number for this, and move over existing users over
to the new function for it.

This also has the side effect of removing the spurious "Unknown IPI
message 1" each time a CPU is awaken.

Signed-off-by: Nicolas Pitre <nico@linaro.org>