ARM: b.L switcher: restore the fundamental switch context assumption
For the switch to be race free, it must happen in a known kernel context.
This means we cannot perform it from an IPI as we don't know what is being
interrupted by that IPI, such as an ongoing cache maintenance loop or any
other operation requiring a stable view of the CPU architecture.

Let's replace usage of smp_call_function_single() by schedule_work_on() to
ensure remote switch requests are queued to be done in process context.
Request to switch the calling CPU is performed immediately as before.

To ensure proper usage, a new
request function is added:
void bL_switch_request(unsigned int cpu, unsigned int new_cluster_id)

Signed-off-by: Nicolas Pitre <nico@linaro.org>