drivers: arm_spc: remove locks for reset upon wfi
This patch is a ultimate hack. Firstly it makes sure the info
pointer is inizialize to avoid resetting a cpu before the spc has
been mapped in Linux.

Secondly, when the cpu reset mechanism is triggered lock must not
be taken since the C bit is cleared and cacheable spinlocks does not
work anymore.