HDLCD: Disable writing to the command register in the hdlcd_disable function.
When running Android on TC2, the screen gets black when the mouse pointer
is moved. It eventually comes back after a couple of seconds and the mouse
pointer has moved. By disabling the write operation to the command register,
the screen gets updated correctly when the mouse pointer moves. It's still
not a smooth interaction but rather flickering.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>