ARM: bL_switcher: move power_down call to the outbound CPU
Let's simplify things a bit.  The calling of power_down on the inbound
CPU is just a remnant of some early attempt to avoid a race where the
inbound CPU would switch back right away and call power_up on the
yet-to-be-down outbound CPU.  To satisfy the proper sequence, power_down
simply has to be called before the inbount CPU is ungated.

Signed-off-by: Nicolas Pitre <nico@linaro.org>