ARM: bL_switcher: enumerate GIC IDs at run time
Currently, GIC IDs are hardcoded making the code dependent on the x4 b.L
configuration.  Let's enumerate GIC IDs upon switcher initialization
to support other b.L configurations such as the x1 one.

To enumerate those IDs, we must perform a switch because the needed
GIC register is banked per CPU.  This switch is obviously done without
any GIC migration as we initially don't know the ID to migrate the GIC to
prior the switch and we switch back immediately once the ID is retrieved.

TODO: make sure the powering off of a CPU in bL_do_switch()does not affect
the not-saved GIC state in the initial CPU.

Signed-off-by: Nicolas Pitre <nico@linaro.org>