ARM: bL_switcher: warn about non valid 'current' pointer when changing stack
This even affects printk() which is not immediately obvious.

This could be worked around by creating a dummy task struct on the
temporary stack.  This would require quite a lot of memory (aligned 8KB)
which is rather wasteful only for the purpose of a valid 'current'.
Let's just note the issue and try to avoid problematic operations for
now.  By default pr_debug() is a no-op anyway.

Signed-off-by: Nicolas Pitre <nico@linaro.org>