ARM: kingfisher: add CPU use counts to the power up/down API implementation
It is possible for an outbound CPU to let its inbound counterpart run
free before it shut itself down, and that inbound CPU might decide to
call power_up() on the yet-to-by-down outbound CPU.  If that power_up()
is followed by an out-of-order power_down(), the CPU won't continue to run
as expected.  Solve this with a simple use count.

Signed-off-by: nicolas Pitre <nico@linaro.org>