ARM: bL_switcher: refactor bL_power_ops for platform coherency exit/setup
This patch makes a few distinct changes, as follows:

  * Split the power_down() method into:

      * bool power_down_prepare(), which nominates the CPU as going
        down, and arbitrates to select a last-man CPU to take down
        the cluster if necessary;

      * power_down(..., bool last_man), which actually powers the
        CPU down, and if necessary the cluster.  This method is
        responsible for doing platform-specific coherency exit
        operations.

        This method is responsible for exiting coherency at all
        levels including L1.  This appears unavoidable, since there
        is no architecturally guaranteed way to exit coherency,
        even at L1.

  * Adds a new platform-specific helper power_up_setup(), which
    does platform-specific setup in preparation for turning the CPU
    on, such as invalidating local caches or entering coherency.

    This method must be assembler for now, since it must run before
    the MMU can be switched on.

  * Provides helper methods to coordinate between CPUs coming down
    and CPUs going up, so that cluster teardown and setup
    operations are not done for a cluster simultaneously:

    For use in power_down():
      * __bL_cpu_going_down(unsigned int cluster, unsigned int cpu)
      * __bL_outbound_enter_critical(unsigned int cluster)
      * __bL_outbound_leave_critical(unsigned int cluster)
      * __bL_cpu_down(unsigned int cluster, unsigned int cpu)

    For use in power_up_setup():
      * .macro __bL_inbound_enter_critical
      * .macro __bL_inbound_leave_critical

L1 teardown is now entirely the responsibility of power_down(),
since there's no generic architectural mapping between coherency
controls, cache levels and power domains.

Unless we actually turn the MMU off, the memory used for the
synchronisation protocol must not be mapped cacheable anywhere,
otherwise we may get unexpected behaviour due to mismatched memory
aliases.

To achieve this without ugliness like statically reseving an
arbitrary range in the DT (which anyway doesn't work since it looks
like such reserved regions get mapped as a variety or Normal Memory
in any case), we grab some memory using memblock_steal().

This can only be done early, during the board's .reserve() method,
so the switcher init code is factored out to include a _reserve()
function which can be called from the board's .reserve() method,
allowing suitable memory to be stolen for this purpose.

Thanks to Nicolas Pitre and Achin Gupta for the help with this
patch.

Signed-off-by: Dave Martin <dave.martin@linaro.org>