ARM: vexpress: Port kingfisher.c to new bL_power_ops and handle CCI
This patch adds CCI-400 (cache-coherent interconnect) snoop
handling to the kingfisher control block platform-specifc code, by
migrating to the new bL_power_ops and adding suitable code to
manipulate the CCI.

The current implementation considers the CCI-400 to be "part of"
the kingfisher control block for now, which is convenient but
untrue.

The CCI support should eventually turn into common code, if other
platforms use it.

This patch also redundantly flushes some cache levels, since
flush_dcache_level() allows a flush _to_ a particular level instead
of PoC, but not _from_ a particular level to PoC (which is what we
may want in this case).

Thanks to Achin Gupta and Nicolas Pitre for their help and
contributions.

Signed-off-by: Dave Martin <dave.martin@linaro.org>