ARM: bL_switcher: Recombine power_down_prepare() and power_down()
Because there is now almost nothing between the calls to the
power_down_prepare() and power_down() methods, it is no longer
really worth them being separate.  This patch combines them.

As a side-effect, this requires the call to __bL_cpu_going_down()
to be made from the platform-specific code, so that needs to be
non-static.

Thanks to Nicolas Pitre for the patch.

Signed-off-by: Dave Martin <dave.martin@linaro.org>