ARM: bL_switcher: improve stack (and mm) isolation
Code called from bL_do_switch() might end up referencing 'current' for
some reasons.  However, 'current' is derived from the stack pointer.
With any arbitrary stack, the returned value for 'current' and any
dereferenced values through it are just random garbage which may lead to
segmentation faults.

The active page table during the execution of bL_do_switch() is also a
problem.  There is no guarantee that the inbound CPU won't destroy the
corresponding task which would free the attached page table while the
outbound CPU is still running and relying on it.

To solve both issues, let's borrow some of the task space belonging to
the init/idle task which, by its nature, is lightly used and therefore is
unlikely to clash with our usage.  The init task is also never going away,
and its mm is used as the master copy for kernel mappings which doesn't
need any lazy paging in the vmalloc area.

Signed-off-by: Nicolas Pitre <nico@linaro.org>