BL IKS on TC2: Quickfixes to make generic switcher work on TC2
ACE & ACP snoops have to be disabled by sw on the TC2 to allow the
STANDBYWFIL2 signal to be asserted. This also disables the coherency
channel and prevents subsequent dsb instructions from completing.
Do not use dsbs for the time being as a result.

Export useful routine to read physical mpidr

Added rudimentary support to measure switching costs through a sysfs
interface