BL IKS on TC2: Use reset status instead of wfi status to determine cluster state
When a cpu is the first on the inbound cluster, it needs t be ensured that
the cluster is powered down before powering it up again. This can be done
either using the STANDBYWFISTAT or CFGREG6 in the SPC/SCC memory map. The
former approach was chosen earlier as its in the SPC memory map but does
not work as A15 deasserts its STANDBYWFI signals when placed in reset by
the M3. Hence using the raw reset status from the CFGREG6 instead.