cpuidle : change sysfs function api
The next patch will use a cpu mask to describe for which
cpu the state is enaled or disabled.

In order to do that, the cpuidle device will be used in the
function. In order to have the correct information, this
patch change the functions to have the cpuidle_device and
the index related to the state and state_usage.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>

Conflicts:

	drivers/cpuidle/sysfs.c