BL IKS on TC2: Workaround to cater for aborted last man operations
An inbound cpu might want to rejoin the outbound cluster after a last
man has been elected on it. 'powerdown_needed' can cause the 'last man'
to abort his ops if the inbound cpu updates the 'cpu_online_map' before
this call is made. Currently there is no way for the inbound cpu to know
that the outbound 'last man' will not power down the cluster. Thinking
its the 'first man' in the cluster, the inbound cpu will end up waiting
infinitely for a cluster shutdown.

This workaround updates the 'cpu_online_map' only after the 'first man'
inbound cpu has been a cluster shutdown. This affects performance as we
are giving up on a chance to abort redundant last man operations.