BL IKS on TC2: Workaround to quell spurious wakeups
It seems that between reading the wfi status of the inbound cpu and
reading the reset status register, it gets woken up spuriously and
arrives at 'bL_entry_point'. Since we are not relying on fake resets
right now, the 'wfe' in 'bL_entry_point' prevents this cpu from being
reset. This patch delays the write to the BX_ADDR register and the
online cpu map till we are sure that the inbound cpu has been placed
in reset.