ARM: TC2: cpuidle hack of the year
This massive hack allows the TC2 idle driver to update the coupled
cpu mask upon IKS CPU switch notification.

It unregisters all existing cpuidle devices, updates the masks and
register them again to leave the CPU idle subsystem in a clean state.

It is provided for back-up in case all hell comes loose and must not
be considered a viable solution.