cpufreq: vexpress: add in-kernel switching(IKS) support
This patch integrates the cpu switching implemented as part of
in-kernel switching feature. This is initial and trivial support
in which the OPPs of both the clusters are merged into single
table(with duplicates).

Whenever the frequencies higher than the maximum LITTLE cluster
frequency is requested, the cpu is switched to the big cluster.
Similarly when the frequencies lower than the minimum big cluster
frequency is requested, the cpu is switched to the LITTLE cluster.

This patch supports only per-cpu switching. It can be extended to
support cluster switching by enforcing the relation between all CPUs
requiring any coordination.

Signed-off-by: Sudeep KarkadaNagesha <sudeep.karkadanagesha@arm.com>