cpufreq: vexpress: add cluster switching support
This patch adds the cluster switching support. This co-exists
with the per-cpu switching and can be enabled dynamically when
needed through debugfs interface.

To enable cluster switching:
$echo 1 > /sys/kernel/debug/cluster_switch/enable
and to disable:
$echo 0 > /sys/kernel/debug/cluster_switch/enable

Since all the per-cpu data needs to be updated whenever the switching
mode is changed from per-cpu to cluster or vice-versa, the cpufreq
driver is unregistered and re-registered back.

Signed-off-by: Sudeep KarkadaNagesha <sudeep.karkadanagesha@arm.com>