BL IKS on TC2: Enable cpu wakeup sources before powering it up.
This patch enables the irq wakeup source for the M3 before powering
up a cpu. Usually each cpu will do this before powering down, but
this is required for the first switch when the inbound cluster will
have been powered down without enabling its wakeup sources.

The same could be done from the bootmon prompt, but that leaves the
inbound cluster susceptible to spurious wakeups when the primary
cpu in linux broadcasts interrupts to wake up its secondaries.