ARM: TC2: port cpuidle to IKS power API
This patch updates the CPU idle driver for TC2 so that it can reuse
the in kernel switcher power API. A switcher notifier is added to cater
for coupled mask in case the integration of IKS and CPU idle fails.

Coupled state flag removed according to the latest firmware updates that
allow per-CPU reset and wake-up capabilities.