cpufreq: vexpress: add hysterisis control to avoid excessive switching
This patch adds the support for hysterisis managment to the cpufreq
driver. The main purpose for the hysterisis control is to avoid excessive
switching that can occur while honouring ondemand frequency requests.
With this control, the frequencies that demand the switch accross the
clusters are tracked and honoured only when the number of similar requests
exceeds the configurable parameter. This parameter can be configured
independently for switches in both direction.

To set this parameter:
$ echo u > /sys/devices/system/cpu/cpu<m>/cpufreq/bl_up_hyst_config
$ echo d > /sys/devices/system/cpu/cpu<m>/cpufreq/bl_down_hyst_config
where: m - cpu, u - up hysterisis value, d - down hysterisis value

Signed-off-by: Sudeep KarkadaNagesha <sudeep.karkadanagesha@arm.com>