ARM: common: bL_switcher fix sync structure init for MP
If the power down API is used for MP, the synchronization structure has
to be initialized for all clusters. This patch loops through the number
of clusters and initialize the structure accordingly.

The horrible #if 1 condition must be replaced with a boot parameter to
distinguish between MP and IKS kernels.