drivers: misc: vexpress: change spinlock to semaphore
The DVFS settings are now interrupt-based and can sleep waiting for
the completion of the DVFS request. But M3 firmware can handle only
one request at a time, enforcing a locking across DVFS requestion.
This patch replaces spinlocks with the semaphores as we need to sleep
holding the lock.

This patch also removes locking couple of unused function. We may need
to address if these functions are used again else the functions can be
removed.

Signed-off-by: Sudeep KarkadaNagesha <sudeep.karkadanagesha@arm.com>