sched: fix arch_get_fast_and_slow_cpus to get logical cpumask correctly
The patch "sched: Use device-tree to provide fast/slow CPU list for HMP"
depends on the ordering of CPU's in the device tree. It breaks to determine
the logical mask correctly if the logical mask of the CPUs differ from
physical ordering in the device tree.

This patch fix the logic by depending on the mpidr in the device tree
and mapping that mpidr to the logical cpu.

Change-Id: I9b3d2136c7ab984b97b5f24908c7dc53cf9750fa
Signed-off-by: Sudeep KarkadaNagesha <sudeep.karkadanagesha@arm.com>