sched: add a new arch_sd_local_flags for sched_domain init
The function arch_sd_local_flags is used to set architecture specific flags in
the sched_domain. A new flag SD_SHARE_POWERDOMAIN is also created to reflect
whether groups of CPUs in a sched_domain level can or not reach a different
power state. If groups of cores can be power gated independently,
as an example, the flag should be cleared at CPU level. This information is
used to decide if it's worth packing some tasks in a group of CPUs in order to
power gated the other groups instead of spreading the tasks. The default
behavior of the scheduler is to spread tasks so the flag is set into all
sched_domains.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>